requirements-dev.txt
pytest
pdfplumber
tabula-py[jpype]
pillow